requirements.txt
numpy~=1.21.5
flask~=2.1.0
wtforms~=3.0.1
werkzeug~=2.0.3
tensorflow-cpu
gunicorn~=20.1.0
pillow~=9.0.1